#50 Added local storage to coding sheets
When the user clicks the "completed" button, this creates a new local storage instance with the name "codingSheet-(CODING SHEET ID)".

When the database is called, it will first check to see if the user has a local storage instance corresponding to the response data. If it does, it updates the list with the provided local storage values.

TODO: Progress bar bugs out because it's not properly hooked up to how the rest of the code works now.